XP-993 Convert Module to an implementation
- Converted Module to an implementation. This implementation does not need a builder since it's only going to be built using one single parameter - Bundle.
- Created a static method named Module.from(Bundle) that will create a Module.
