+++ modules/core-api/src/test/java/com/enonic/xp/module/ModuleTest.java
@@ -1,25 +1,26 @@
-package com.enonic.xp.core.impl.module;
+package com.enonic.xp.module;
 
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Collections;
+import java.util.Dictionary;
 import java.util.Enumeration;
+import java.util.Hashtable;
 import java.util.List;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.osgi.framework.Bundle;
+import org.osgi.framework.Constants;
+import org.osgi.framework.Version;
 
 import com.google.common.collect.Lists;
-
-import com.enonic.xp.module.ModuleKey;
-import com.enonic.xp.module.ModuleVersion;
 
 import static org.junit.Assert.*;
 import static org.mockito.Matchers.isA;
 
-public class ModuleImplTest
+public class ModuleTest
 {
     private Bundle bundle;
 
@@ -33,19 +34,13 @@ public class ModuleImplTest
     @Test
     public void testCreateModule()
     {
-        final ModuleImpl module = new ModuleImpl();
-        module.moduleKey = ModuleKey.from( "mymodule" );
-        module.moduleVersion = ModuleVersion.from( "1.0.0" );
-        module.displayName = "module display name";
-        module.url = "http://enonic.net";
-        module.vendorName = "Enonic";
-        module.vendorUrl = "https://www.enonic.com";
+        final Module module = Module.from( bundle );
 
         assertEquals( "mymodule", module.getKey().toString() );
-        assertEquals( "module display name", module.getDisplayName() );
-        assertEquals( "http://enonic.net", module.getUrl() );
-        assertEquals( "Enonic", module.getVendorName() );
-        assertEquals( "https://www.enonic.com", module.getVendorUrl() );
+        assertEquals( "mymodule", module.getDisplayName() );
+        assertEquals( "http://enonic.com/path/to/module", module.getUrl() );
+        assertEquals( "Enonic AS", module.getVendorName() );
+        assertEquals( "http://enonic.com", module.getVendorUrl() );
     }
 
     private Bundle mockBundle( final String... resourcePaths )
@@ -68,6 +63,19 @@ public class ModuleImplTest
         final Enumeration<URL> bundleEntries = Collections.enumeration( urlList );
         Mockito.when( bundle.findEntries( isA( String.class ), isA( String.class ), isA( Boolean.class ) ) ).thenReturn( bundleEntries );
         Mockito.when( bundle.getState() ).thenReturn( Bundle.ACTIVE );
+        Mockito.when( bundle.getSymbolicName() ).thenReturn( "mymodule" );
+        Mockito.when( bundle.getVersion() ).thenReturn( new Version( 1, 2, 0 ) );
+        Mockito.when( bundle.getHeaders() ).thenReturn( createBundleHeaders() );
         return bundle;
+    }
+
+    private Dictionary<String, String> createBundleHeaders() {
+        Dictionary<String, String> headers = new Hashtable<String, String>(  );
+        headers.put( Constants.BUNDLE_NAME, "mymodule" );
+        headers.put( Module.X_MODULE_URL, "http://enonic.com/path/to/module" );
+        headers.put( Module.X_VENDOR_NAME, "Enonic AS" );
+        headers.put( Module.X_VENDOR_URL, "http://enonic.com" );
+
+        return headers;
     }
 }
