+++ modules/core-module/src/main/java/com/enonic/xp/core/impl/module/ModuleRegistryImpl.java
@@ -87,7 +87,7 @@ public final class ModuleRegistryImpl
 
     private boolean isModule( final Bundle bundle )
     {
-        return ( bundle.getState() != Bundle.UNINSTALLED ) && ModuleBuilder.isModule( bundle );
+        return ( bundle.getState() != Bundle.UNINSTALLED ) && Module.isModule( bundle );
     }
 
     private void publishModuleChangeEvent( final BundleEvent event )
@@ -117,10 +117,7 @@ public final class ModuleRegistryImpl
 
     private void installModule( final Bundle bundle )
     {
-        final ModuleBuilder builder = new ModuleBuilder();
-        builder.bundle( bundle );
-
-        final Module module = builder.build();
+        final Module module = Module.from( bundle );
         installModule( module );
     }
