+++ modules/core-module/src/test/java/com/enonic/xp/core/impl/module/ModuleServiceImplTest.java
@@ -1,11 +1,21 @@
 package com.enonic.xp.core.impl.module;
 
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.time.Instant;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Dictionary;
+import java.util.Enumeration;
+import java.util.Hashtable;
+import java.util.List;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.osgi.framework.Bundle;
+import org.osgi.framework.Constants;
+import org.osgi.framework.Version;
 
 import com.google.common.collect.Lists;
 
@@ -13,10 +23,10 @@ import com.enonic.xp.module.Module;
 import com.enonic.xp.module.ModuleKey;
 import com.enonic.xp.module.ModuleKeys;
 import com.enonic.xp.module.ModuleNotFoundException;
-import com.enonic.xp.module.ModuleVersion;
 import com.enonic.xp.module.Modules;
 
 import static org.junit.Assert.*;
+import static org.mockito.Matchers.isA;
 
 public class ModuleServiceImplTest
 {
@@ -34,19 +44,7 @@ public class ModuleServiceImplTest
 
     private Module createModule( final String key )
     {
-        final ModuleImpl module = new ModuleImpl();
-        module.moduleKey = ModuleKey.from( key );
-        module.moduleVersion = ModuleVersion.from( "1.0.0" );
-        module.displayName = "module display name";
-        module.url = "http://enonic.net";
-        module.vendorName = "Enonic";
-        module.vendorUrl = "https://www.enonic.com";
-
-        final Bundle bundle = Mockito.mock( Bundle.class );
-        Mockito.when( bundle.getState() ).thenReturn( Bundle.ACTIVE );
-        Mockito.when( bundle.getLastModified() ).thenReturn( Instant.parse( "2012-01-01T00:00:00.00Z" ).toEpochMilli() );
-        module.bundle = bundle;
-
+        final Module module = Module.from( mockBundle( key ) );
         return module;
     }
 
@@ -110,5 +108,43 @@ public class ModuleServiceImplTest
 
         this.service.stopModule( module.getKey() );
         Mockito.verify( module.getBundle() ).stop();
+    }
+
+    private Bundle mockBundle( final String key )
+    {
+        List<String> resourcePaths = Arrays.asList( "app/site.xml", "app/parts/mypart/part.xml", "app/pages/mypage/page.xml" );
+        final Bundle bundle = Mockito.mock( Bundle.class );
+        final List<URL> urlList = Lists.newArrayList();
+        for ( String resourcePath : resourcePaths )
+        {
+            try
+            {
+                final URL url = new URL( "http://109.0:1/" + resourcePath );
+                urlList.add( url );
+                Mockito.when( bundle.getResource( resourcePath ) ).thenReturn( url );
+            }
+            catch ( MalformedURLException e )
+            {
+                throw new RuntimeException( e );
+            }
+        }
+        final Enumeration<URL> bundleEntries = Collections.enumeration( urlList );
+        Mockito.when( bundle.findEntries( isA( String.class ), isA( String.class ), isA( Boolean.class ) ) ).thenReturn( bundleEntries );
+        Mockito.when( bundle.getState() ).thenReturn( Bundle.ACTIVE );
+        Mockito.when( bundle.getSymbolicName() ).thenReturn( key );
+        Mockito.when( bundle.getVersion() ).thenReturn( new Version( 1, 2, 0 ) );
+        Mockito.when( bundle.getHeaders() ).thenReturn( createBundleHeaders() );
+        Mockito.when( bundle.getLastModified() ).thenReturn( Instant.parse( "2012-01-01T00:00:00.00Z" ).toEpochMilli() );
+        return bundle;
+    }
+
+    private Dictionary<String, String> createBundleHeaders() {
+        Dictionary<String, String> headers = new Hashtable<String, String>(  );
+        headers.put( Constants.BUNDLE_NAME, "mymodule" );
+        headers.put( Module.X_MODULE_URL, "http://enonic.com/path/to/module" );
+        headers.put( Module.X_VENDOR_NAME, "Enonic AS" );
+        headers.put( Module.X_VENDOR_URL, "http://enonic.com" );
+
+        return headers;
     }
 }
